Update to move pdb for debugging algos

+++ run_docker.sh
@@ -21,6 +21,7 @@ default_results_dir=$current_dir
 default_config_file=$current_dir/Launcher/config.json
 python_location=$current_dir/Algorithm.Python/
 csharp_dll=$current_dir/Launcher/bin/Debug/QuantConnect.Algorithm.CSharp.dll
+csharp_pdb=$current_dir/Launcher/bin/Debug/QuantConnect.Algorithm.CSharp.pdb
 
 if [ -f "$1" ]; then
     IFS="="
@@ -65,10 +66,16 @@ if [ ! -d "$results_dir" ]; then
     exit 1
 fi
 
+if [ ! -f "$csharp_dll" ]; then
+    echo "Csharp dll at $csharp_dll does not exist; be sure to build the project first"
+    exit 1
+fi
+
 docker run --rm --mount type=bind,source=$config_file,target=/Lean/Launcher/config.json,readonly \
     --mount type=bind,source=$data_dir,target=/Data,readonly \
     --mount type=bind,source=$results_dir,target=/Results \
     --mount type=bind,source=$csharp_dll,target=/Lean/Launcher/bin/Debug/QuantConnect.Algorithm.CSharp.dll \
+    --mount type=bind,source=$csharp_pdb,target=/Lean/Launcher/bin/Debug/QuantConnect.Algorithm.CSharp.pdb \
     --mount type=bind,source=$python_location,target=/Lean/Algorithm.Python \
     -p 55555:55555 -p 5678:5678 \
     --name LeanEngine \
